Language.Ninja.AST

Description

The Ninja build language after parsing.

This module re-exports all of the modules under the Language.Ninja.AST namespace for convenience.

It is recommended that you import it with the following style:

import qualified Language.Ninja.AST as AST

Since: 0.1.0

Language.Ninja.AST.Ninja

data Ninja ann

A parsed Ninja file.

Since: 0.1.0

makeNinja :: Monoid ann => Ninja ann

Construct a Ninja with all default values

Since: 0.1.0

ninjaRules :: Lens' (Ninja ann) (HashMap Text (Rule ann))

The rules defined in a parsed Ninja file.

Since: 0.1.0

ninjaSingles :: Lens' (Ninja ann) (HashMap Text (Build ann))

The set of build declarations with precisely one output.

Since: 0.1.0

ninjaMultiples :: Lens' (Ninja ann) (HashMap Outputs (Build ann))

The set of build declarations with two or more outputs.

Since: 0.1.0

ninjaPhonys :: Lens' (Ninja ann) (HashMap Text (HashSet Text))

The set of phony build declarations.

Since: 0.1.0

ninjaDefaults :: Lens' (Ninja ann) (HashSet Text)

The set of default targets.

Since: 0.1.0

ninjaPools :: Lens' (Ninja ann) (HashMap Text Int)

A mapping from pool names to pool depth integers.

Since: 0.1.0

ninjaSpecials :: Lens' (Ninja ann) (HashMap Text Text)

A map from "special" top-level variables to their values.

Since: 0.1.0

Language.Ninja.AST.Build

data Build ann

A parsed Ninja build declaration.

Since: 0.1.0

makeBuild

Arguments

:: Monoid ann

=> Text
The rule name

-> Env Text Text
The environment

-> Build ann

Construct a Build with all default values.

Since: 0.1.0

buildRule :: Lens' (Build ann) Text

A lens into the rule name associated with a Build.

Since: 0.1.0

buildEnv :: Lens' (Build ann) (Env Text Text)

A lens into the environment associated with a Build.

Since: 0.1.0

buildDeps :: Lens' (Build ann) (Deps ann)

A lens into the dependencies associated with a Build.

Since: 0.1.0

buildBind :: Lens' (Build ann) (HashMap Text Text)

A lens into the bindings associated with a Build.

Since: 0.1.0

Language.Ninja.AST.Deps

data Deps ann

A set of Ninja build dependencies.

Since: 0.1.0

makeDeps :: Monoid ann => Deps ann

Construct a Deps with all default values

Since: 0.1.0

depsNormal :: Lens' (Deps ann) (HashSet Text)

A lens into the set of normal dependencies in a Deps.

Since: 0.1.0

depsImplicit :: Lens' (Deps ann) (HashSet Text)

A lens into the set of implicit dependencies in a Deps.

Since: 0.1.0

depsOrderOnly :: Lens' (Deps ann) (HashSet Text)

A lens into the set of order-only dependencies in a Deps.

Since: 0.1.0

Language.Ninja.AST.Rule

data Rule ann

A parsed Ninja rule declaration.

Since: 0.1.0

makeRule :: Monoid ann => Rule ann

Construct a Rule with all default values

Since: 0.1.0

ruleBind :: Lens' (Rule ann) (HashMap Text (Expr ann))

The set of bindings in scope during the execution of this rule.

Since: 0.1.0

Language.Ninja.AST.Expr

data Expr ann

An expression containing variable references in the Ninja language.

Since: 0.1.0

Constructors

Exprs !ann ![Expr ann]
Sequencing of expressions.

Since: 0.1.0

Lit !ann !Text
A literal string.

Since: 0.1.0

Var !ann !Text
A variable reference.

Since: 0.1.0

_Exprs :: Prism' (Expr ann) (ann, [Expr ann])

A prism for the Exprs constructor.

Since: 0.1.0

_Lit :: Prism' (Expr ann) (ann, Text)

A prism for the Lit constructor.

Since: 0.1.0

_Var :: Prism' (Expr ann) (ann, Text)

A prism for the Var constructor.

Since: 0.1.0

askVar :: Env Text Text -> Text -> Text

Look up the given variable in the given context, returning the empty string if the variable was not found.

Since: 0.1.0

askExpr :: Env Text Text -> Expr ann -> Text

Evaluate the given Expr in the given context (Env Text Text).

Since: 0.1.0

addBind :: Text -> Expr ann -> Env Text Text -> Env Text Text

Add a binding with the given name (Text) and value (Expr) to the given context.

Since: 0.1.0

addBinds :: [(Text, Expr ann)] -> Env Text Text -> Env Text Text

Add bindings from a list. Note that this function evaluates all the right-hand-sides first, and then adds them all to the environment.

For example:

>>> let binds = [("x", Lit "5"), ("y", Var "x")]
>>> AST.headEnv (addBinds binds AST.makeEnv)
fromList [("x","5"),("y","")]

Since: 0.1.0

normalizeExpr :: forall ann. Monoid ann => Expr ann -> Expr ann

Normalize an Expr by recursively flattening any Exprs nodes, removing empty Lit nodes, combining adjacent Lit nodes, and pulling out the interior of the top-level Exprs node if it has only one subnode.

The number of Exprs nodes in the output is guaranteed to be 0 or 1.

If it is 0, then there is exactly one node of any type in the output.

The output is thus isomorphic to (Maybe ann, [(ann, Either Text Text)]), where the Maybe ann represents the annotation of the top-level Exprs node if it exists.

Since: 0.1.0

Language.Ninja.AST.Env

data Env k v

A Ninja-style environment, basically a nonempty list of hash tables.

Since: 0.1.0

makeEnv :: Env k v

Construct an empty environment.

Since: 0.1.0

fromEnv :: Iso' (Env k v) (Maps k v)

An isomorphism between an Env and a nonempty list of HashMaps.

Since: 0.1.0

addEnv :: Key k => k -> v -> Env k v -> Env k v

Add the given key and value to the given Env.

Since: 0.1.0

askEnv :: Key k => Env k v -> k -> Maybe v

Look up the given key in the given Env.

Since: 0.1.0

scopeEnv :: Env k v -> Env k v

Push a new Env onto the stack.

Since: 0.1.0