Language.Ninja.AST.Build

Description

This module contains a type representing a build declaration in the parsed Ninja AST, along with any supporting or related types.

Since: 0.1.0

Build

data Build ann

A parsed Ninja build declaration.

Since: 0.1.0

makeBuild

Arguments

:: Monoid ann

=> Text
The rule name

-> Env Text Text
The environment

-> Build ann

Construct a Build with all default values.

Since: 0.1.0

buildRule :: Lens' (Build ann) Text

A lens into the rule name associated with a Build.

Since: 0.1.0

buildEnv :: Lens' (Build ann) (Env Text Text)

A lens into the environment associated with a Build.

Since: 0.1.0

buildDeps :: Lens' (Build ann) (Deps ann)

A lens into the dependencies associated with a Build.

Since: 0.1.0

buildBind :: Lens' (Build ann) (HashMap Text Text)

A lens into the bindings associated with a Build.

Since: 0.1.0

type BuildConstraint c ann = (DepsConstraint c ann, c Text, c (HashMap Text Text), c (Maps Text Text), c ann)

The set of constraints required for a given constraint to be automatically computed for a Build.

Since: 0.1.0