Language.Ninja.AST.Deps

Description

This module contains a type representing the dependencies of a build declaration in the parsed Ninja AST, along with any supporting or related types.

Since: 0.1.0

Deps

data Deps ann

A set of Ninja build dependencies.

Since: 0.1.0

makeDeps :: Monoid ann => Deps ann

Construct a Deps with all default values

Since: 0.1.0

depsNormal :: Lens' (Deps ann) (HashSet Text)

A lens into the set of normal dependencies in a Deps.

Since: 0.1.0

depsImplicit :: Lens' (Deps ann) (HashSet Text)

A lens into the set of implicit dependencies in a Deps.

Since: 0.1.0

depsOrderOnly :: Lens' (Deps ann) (HashSet Text)

A lens into the set of order-only dependencies in a Deps.

Since: 0.1.0

type DepsConstraint c ann = (c (HashSet Text), c ann)

The set of constraints required for a given constraint to be automatically computed for a Ninja.

Since: 0.1.0