Language.Ninja.AST.Env

Description

This module contains a type representing a Ninja-style environment along with any supporting or related types.

Since: 0.1.0

Documentation

data Env k v

A Ninja-style environment, basically a nonempty list of hash tables.

Since: 0.1.0

makeEnv :: Env k v

Construct an empty environment.

Since: 0.1.0

fromEnv :: Iso' (Env k v) (Maps k v)

An isomorphism between an Env and a nonempty list of HashMaps.

Since: 0.1.0

headEnv :: Env k v -> HashMap k v

Get the first HashMap in the underlying nonempty list.

Since: 0.1.0

tailEnv :: Env k v -> Maybe (Env k v)

If the remainder of the underlying nonempty list is nonempty, return the remainder after Env wrapping. Otherwise, return Nothing.

Since: 0.1.0

scopeEnv :: Env k v -> Env k v

Push a new Env onto the stack.

Since: 0.1.0

addEnv :: Key k => k -> v -> Env k v -> Env k v

Add the given key and value to the given Env.

Since: 0.1.0

askEnv :: Key k => Env k v -> k -> Maybe v

Look up the given key in the given Env.

Since: 0.1.0

type EnvConstraint c k v = (Key k, c k, c v, c (Maps k v))

The set of constraints required for a given constraint to be automatically computed for an Env.

Since: 0.1.0

type Key k = (Eq k, Hashable k)

A constraint alias for (Eq k, Hashable k).

Since: 0.1.0

type Maps k v = NonEmpty (HashMap k v)

A NonEmpty list of HashMaps.

Since: 0.1.0