Language.Ninja.AST.Expr

Description

This module contains a type representing a string that potentially contains variable references in the parsed Ninja AST, along with any supporting or related types.

Since: 0.1.0

Expr

data Expr ann

An expression containing variable references in the Ninja language.

Since: 0.1.0

Constructors

Exprs !ann ![Expr ann]
Sequencing of expressions.

Since: 0.1.0

Lit !ann !Text
A literal string.

Since: 0.1.0

Var !ann !Text
A variable reference.

Since: 0.1.0

_Exprs :: Prism' (Expr ann) (ann, [Expr ann])

A prism for the Exprs constructor.

Since: 0.1.0

_Lit :: Prism' (Expr ann) (ann, Text)

A prism for the Lit constructor.

Since: 0.1.0

_Var :: Prism' (Expr ann) (ann, Text)

A prism for the Var constructor.

Since: 0.1.0

askVar :: Env Text Text -> Text -> Text

Look up the given variable in the given context, returning the empty string if the variable was not found.

Since: 0.1.0

askExpr :: Env Text Text -> Expr ann -> Text

Evaluate the given Expr in the given context (Env Text Text).

Since: 0.1.0

addBind :: Text -> Expr ann -> Env Text Text -> Env Text Text

Add a binding with the given name (Text) and value (Expr) to the given context.

Since: 0.1.0

addBinds :: [(Text, Expr ann)] -> Env Text Text -> Env Text Text

Add bindings from a list. Note that this function evaluates all the right-hand-sides first, and then adds them all to the environment.

For example:

>>> let binds = [("x", Lit "5"), ("y", Var "x")]
>>> AST.headEnv (addBinds binds AST.makeEnv)
fromList [("x","5"),("y","")]

Since: 0.1.0

normalizeExpr :: forall ann. Monoid ann => Expr ann -> Expr ann

Normalize an Expr by recursively flattening any Exprs nodes, removing empty Lit nodes, combining adjacent Lit nodes, and pulling out the interior of the top-level Exprs node if it has only one subnode.

The number of Exprs nodes in the output is guaranteed to be 0 or 1.

If it is 0, then there is exactly one node of any type in the output.

The output is thus isomorphic to (Maybe ann, [(ann, Either Text Text)]), where the Maybe ann represents the annotation of the top-level Exprs node if it exists.

Since: 0.1.0

type ExprConstraint c ann = (c Text, c ann)

The set of constraints required for a given constraint to be automatically computed for a Expr.

Since: 0.1.0