Language.Ninja.AST.Ninja

Description

This module contains a type representing a parsed Ninja file AST, along with any supporting or related types.

Since: 0.1.0

Ninja

data Ninja ann

A parsed Ninja file.

Since: 0.1.0

makeNinja :: Monoid ann => Ninja ann

Construct a Ninja with all default values

Since: 0.1.0

ninjaRules :: Lens' (Ninja ann) (HashMap Text (Rule ann))

The rules defined in a parsed Ninja file.

Since: 0.1.0

ninjaSingles :: Lens' (Ninja ann) (HashMap Text (Build ann))

The set of build declarations with precisely one output.

Since: 0.1.0

ninjaMultiples :: Lens' (Ninja ann) (HashMap Outputs (Build ann))

The set of build declarations with two or more outputs.

Since: 0.1.0

ninjaPhonys :: Lens' (Ninja ann) (HashMap Text (HashSet Text))

The set of phony build declarations.

Since: 0.1.0

ninjaDefaults :: Lens' (Ninja ann) (HashSet Text)

The set of default targets.

Since: 0.1.0

ninjaPools :: Lens' (Ninja ann) (HashMap Text Int)

A mapping from pool names to pool depth integers.

Since: 0.1.0

ninjaSpecials :: Lens' (Ninja ann) (HashMap Text Text)

A map from "special" top-level variables to their values.

Since: 0.1.0

type NinjaConstraint c ann = (BuildConstraint c ann, c (HashMap (HashSet Text) (Build ann)), c (HashMap Text (HashSet Text)), c (HashMap Text (Rule ann)), c (HashMap Text (Build ann)), c (HashMap Text Int), c ann)

The set of constraints required for a given constraint to be automatically computed for a Ninja.

Since: 0.1.0

type Outputs = HashSet Output

A type representing the set of outputs for a build declaration with multiple outputs.

Since: 0.1.0

type Output = Text

A type representing an output of a build declaration.

Since: 0.1.0