Language.Ninja.Compile

Description

Compile a parsed Ninja file into an intermediate representation.

Since: 0.1.0

Documentation

compile :: forall m ann. MonadError CompileError m => Ninja ann -> m Ninja

Compile an parsed Ninja file into a intermediate representation.

Since: 0.1.0