Language.Ninja.Errors

Description

This module re-exports all of the modules under the Language.Ninja.Errors namespace for convenience.

It is recommended that you import it with the following style:

import qualified Language.Ninja.Errors as Errors

Since: 0.1.0

Language.Ninja.Errors

data NinjaError

This type subsumes any error that can be thrown during execution of a function defined in language-ninja.

Since: 0.1.0

Constructors

GenericNinjaError !Text
Generic catch-all error constructor. Avoid using this.

Since: 0.1.0

NinjaParseError !ParseError
Errors encountered during parsing.

Since: 0.1.0

NinjaCompileError !CompileError
Errors encountered during compilation.

Since: 0.1.0

throwNinjaError :: MonadError NinjaError m => NinjaError -> m a

Throw a NinjaError.

Since: 0.1.0

throwGenericNinjaError :: MonadError NinjaError m => Text -> m a

Throw a generic catch-all NinjaError.

Since: 0.1.0

throwNinjaParseError :: MonadError NinjaError m => ParseError -> m a

Throw a ParseError.

Since: 0.1.0

throwNinjaCompileError :: MonadError NinjaError m => CompileError -> m a

Throw a CompileError.

Since: 0.1.0

Language.Ninja.Errors.Compile

data CompileError

The type of errors encountered during compilation.

Since: 0.1.0

Constructors

GenericCompileError !Text
Generic catch-all error constructor. Avoid using this.

Since: 0.1.0

CompileMetaError !CompileMetaError
Errors encountered while compiling a Meta.

Since: 0.1.0

CompileBuildError !CompileBuildError
Errors encountered while compiling a Build.

Since: 0.1.0

CompileRuleError !CompileRuleError
Errors encountered while compiling a Rule.

Since: 0.1.0

CompilePhonyError !CompilePhonyError
Errors encountered while compiling the phony HashMap.

Since: 0.1.0

CompileDefaultError !CompileDefaultError
Errors encountered while compiling the default target HashSet.

Since: 0.1.0

CompilePoolError !CompilePoolError
Errors encountered while compiling a Pool.

Since: 0.1.0

throwCompileError :: MonadError CompileError m => CompileError -> m a

Throw a CompileError.

Since: 0.1.0

throwGenericCompileError :: MonadError CompileError m => Text -> m a

Throw a generic catch-all CompileError.

Since: 0.1.0

data CompileMetaError

The type of errors encountered while compiling Ninja metadata.

Since: 0.1.0

Constructors

GenericCompileMetaError !Text
Generic catch-all error constructor. Avoid using this.

Since: 0.1.0

VersionParseFailure !ParsingError
Failed to parse ninja_required_version: …

Since: 0.1.0

throwCompileMetaError :: MonadError CompileError m => CompileMetaError -> m a

Throw a CompileMetaError.

Since: 0.1.0

throwGenericCompileMetaError :: MonadError CompileError m => Text -> m a

Throw a generic catch-all CompileMetaError.

Since: 0.1.0

throwVersionParseFailure :: MonadError CompileError m => ParsingError -> m a

Throw a VersionParseFailure error.

Since: 0.1.0

data CompilePhonyError

The type of errors encountered while compiling a Ninja phony build.

Since: 0.1.0

Constructors

GenericCompilePhonyError !Text
Generic catch-all error constructor. Avoid using this.

Since: 0.1.0

throwCompilePhonyError :: MonadError CompileError m => CompilePhonyError -> m a

Throw a CompilePhonyError.

Since: 0.1.0

throwGenericCompilePhonyError :: MonadError CompileError m => Text -> m a

Throw a generic catch-all CompilePhonyError.

Since: 0.1.0

data CompileDefaultError

The type of errors encountered while compiling a Ninja default statement.

Since: 0.1.0

Constructors

GenericCompileDefaultError !Text
Generic catch-all error constructor. Avoid using this.

Since: 0.1.0

throwCompileDefaultError :: MonadError CompileError m => CompileDefaultError -> m a

Throw a CompileDefaultError.

Since: 0.1.0

throwGenericCompileDefaultError :: MonadError CompileError m => Text -> m a

Throw a generic catch-all CompileDefaultError.

Since: 0.1.0

data CompileBuildError

The type of errors encountered while compiling a Ninja build statement.

Since: 0.1.0

Constructors

GenericCompileBuildError !Text
Generic catch-all error constructor. Avoid using this.

Since: 0.1.0

BuildRuleNotFound !Text
Rule not found: text

Since: 0.1.0

throwCompileBuildError :: MonadError CompileError m => CompileBuildError -> m a

Throw a CompileBuildError.

Since: 0.1.0

throwGenericCompileBuildError :: MonadError CompileError m => Text -> m a

Throw a generic catch-all CompileBuildError.

Since: 0.1.0

throwBuildRuleNotFound :: MonadError CompileError m => Text -> m a

Throw a BuildRuleNotFound error.

Since: 0.1.0

data CompileRuleError

The type of errors encountered while compiling a Ninja rule statement.

Since: 0.1.0

Constructors

GenericCompileRuleError !Text
Generic catch-all error constructor. Avoid using this.

Since: 0.1.0

RuleLookupFailure !Text
Lookup failed on rule variable: text

Since: 0.1.0

UnknownDepsValue !Text
Unknown deps value: text

Since: 0.1.0

UnexpectedMSVCPrefix !Text
Unexpected msvc_deps_prefix for `deps = "text"`

Since: 0.1.0

throwCompileRuleError :: MonadError CompileError m => CompileRuleError -> m a

Throw a CompileRuleError.

Since: 0.1.0

throwGenericCompileRuleError :: MonadError CompileError m => Text -> m a

Throw a generic catch-all CompileRuleError.

Since: 0.1.0

throwRuleLookupFailure :: MonadError CompileError m => Text -> m a

Throw a RuleLookupFailure error.

Since: 0.1.0

throwUnknownDeps :: MonadError CompileError m => Text -> m a

Throw an UnknownDeps error.

Since: 0.1.0

throwUnexpectedMSVCPrefix :: MonadError CompileError m => Text -> m a

Throw an UnexpectedMSVCPrefix error.

Since: 0.1.0

data CompilePoolError

The type of errors encountered while compiling a Ninja pool statement.

Since: 0.1.0

Constructors

GenericCompilePoolError !Text
Generic catch-all error constructor. Avoid using this.

Since: 0.1.0

InvalidPoolDepth !Int
Invalid pool depth for console: int

Since: 0.1.0

EmptyPoolName
Pool name is an empty string

Since: 0.1.0

throwCompilePoolError :: MonadError CompileError m => CompilePoolError -> m a

Throw a CompilePoolError.

Since: 0.1.0

throwGenericCompilePoolError :: MonadError CompileError m => Text -> m a

Throw a generic catch-all CompilePoolError.

Since: 0.1.0

throwInvalidPoolDepth :: MonadError CompileError m => Int -> m a

Throw an InvalidPoolDepth error.

Since: 0.1.0

throwEmptyPoolName :: MonadError CompileError m => m a

Throw an EmptyPoolName error.

Since: 0.1.0

Language.Ninja.Errors.Parser

data ParseError

The type of errors encountered during parsing.

Since: 0.1.0

Constructors

GenericParseError !Text
Generic catch-all error constructor. Avoid using this.

Since: 0.1.0

LexBindingFailure !Text
Lexer failed at binding: text

Since: 0.1.0

LexExpectedColon
Expected a colon

Since: 0.1.0

LexUnexpectedDollar
Unexpected $ followed by unexpected stuff

Since: 0.1.0

LexUnexpectedSeparator Char
Lexer expected a separator character but found something else

Since: 0.1.0

LexParsecError !(ParseError Char Dec)
Any other lexer error.

Since: 0.1.0

ParseBadDepthField !Text
Could not parse depth field in pool, got: text

Since: 0.1.0

ParseUnexpectedBinding !Text
Unexpected binding defining text

Since: 0.1.0

throwParseError :: MonadError ParseError m => ParseError -> m a

Throw a ParseError.

Since: 0.1.0

throwGenericParseError :: MonadError ParseError m => Text -> m a

Throw a generic catch-all ParseError.

Since: 0.1.0

throwLexBindingFailure :: MonadError ParseError m => Text -> m a

Throw a LexBindingFailure error.

Since: 0.1.0

throwLexExpectedColon :: MonadError ParseError m => m a

Throw a LexExpectedColon error.

Since: 0.1.0

throwLexUnexpectedDollar :: MonadError ParseError m => m a

Throw a LexUnexpectedColon error.

Since: 0.1.0

throwLexUnexpectedSeparator :: MonadError ParseError m => Char -> m a

Throw a LexUnexpectedSeparator error.

Since: 0.1.0

throwLexParsecError :: MonadError ParseError m => ParseError Char Dec -> m a

Throw a LexParsecError error.

Since: 0.1.0

throwParseBadDepthField :: MonadError ParseError m => Text -> m a

Throw a ParseBadDepthField error.

Since: 0.1.0

throwParseUnexpectedBinding :: MonadError ParseError m => Text -> m a

Throw a ParseUnexpectedBinding error.

Since: 0.1.0