Language.Ninja.IR

Description

The Ninja build language intermediate representation after compilation.

This module re-exports all of the modules under the Language.Ninja.IR namespace for convenience.

It is recommended that you import it with the following style:

import qualified Language.Ninja.IR as IR

Since: 0.1.0

Language.Ninja.IR.Ninja

data Ninja

A parsed and normalized Ninja file.

Since: 0.1.0

makeNinja :: Ninja

Construct a default Ninja value.

Since: 0.1.0

ninjaMeta :: Lens' Ninja Meta

Metadata, which includes top-level variables like builddir.

Since: 0.1.0

ninjaBuilds :: Lens' Ninja (HashSet Build)

Compiled build declarations.

Since: 0.1.0

ninjaPhonys :: Lens' Ninja (HashMap Target (HashSet Target))

Phony targets, as documented here.

Since: 0.1.0

ninjaDefaults :: Lens' Ninja (HashSet Target)

The set of default targets, as documented here.

Since: 0.1.0

ninjaPools :: Lens' Ninja (HashSet Pool)

The set of pools for this Ninja file.

Since: 0.1.0

Language.Ninja.IR.Meta

data Meta

Ninja top-level metadata, as documented here.

Since: 0.1.0

makeMeta :: Meta

Construct a default Meta value.

Since: 0.1.0

metaReqVersion :: Lens' Meta (Maybe Version)

Corresponds to the ninja_required_version top-level variable.

Since: 0.1.0

metaBuildDir :: Lens' Meta (Maybe Path)

Corresponds to the builddir top-level variable.

Since: 0.1.0

Language.Ninja.IR.Build

data Build

A Ninja build declaration, as documented here.

Since: 0.1.0

makeBuild :: Rule -> Build

Construct a default Build from the given Rule

Since: 0.1.0

buildRule :: Lens' Build Rule

The rule to execute when building any of the outputs.

Since: 0.1.0

buildOuts :: Lens' Build (HashSet Output)

The outputs that are built as a result of rule execution.

Since: 0.1.0

buildDeps :: Lens' Build (HashSet Dependency)

The dependencies that must be satisfied before this can be built.

Since: 0.1.0

Language.Ninja.IR.Rule

data Rule

A Ninja rule declaration, as documented here.

Since: 0.1.0

makeRule

Arguments

:: Text
The rule name.

-> Command
The command to run.

-> Rule
A rule that runs this command.

Construct an Rule with the given name and command, with default values for all other attributes (e.g.: False, Nothing, poolDefault).

Since: 0.1.0

ruleName :: Lens' Rule Text

The name of the rule.

Since: 0.1.0

ruleCommand :: Lens' Rule Command

The command that this rule will run.

Since: 0.1.0

ruleDescription :: Lens' Rule (Maybe Text)

A short description of the command, used to pretty-print the command as it's running. The ninja -v flag controls whether to print the full command or its description; if a command fails, the full command line will always be printed before the command's output.

Since: 0.1.0

rulePool :: Lens' Rule PoolName

The process pool in which this rule will be executed.

Since: 0.1.0

ruleDepfile :: Lens' Rule (Maybe Path)

If set, this should be a path to an optional Makefile that contains extra implicit dependencies. This is used to support C/C++ header dependencies. For more information, read the Ninja documentation here.

Since: 0.1.0

ruleSpecialDeps :: Lens' Rule (Maybe SpecialDeps)

If set, enables special dependency processing used in C/C++ header dependencies. For more information, read the Ninja documentation here.

Since: 0.1.0

ruleGenerator :: Lens' Rule Bool

If this is true, specifies that this rule is used to re-invoke the generator program. Files built using generator rules are treated specially in two ways: firstly, they will not be rebuilt if the command line changes; and secondly, they are not cleaned by default.

Since: 0.1.0

ruleRestat :: Lens' Rule Bool

If true, causes Ninja to re-stat the command's outputs after execution of the command. Each output whose modification time the command did not change will be treated as though it had never needed to be built. This may cause the output's reverse dependencies to be removed from the list of pending build actions.

Since: 0.1.0

ruleResponseFile :: Lens' Rule (Maybe ResponseFile)

If present, Ninja will use a response file for the given command, i.e. write the selected string to the given file before calling the command and delete the file after the command is done.

This is particularly useful on Windows OS, where the maximal length of a command line is limited and response files must be used instead.

Since: 0.1.0

data SpecialDeps

Special dependency information, as described here.

Since: 0.1.0

makeSpecialDepsGCC :: SpecialDeps

Construct a SpecialDeps corresponding to the case in which deps = gcc is set in a Ninja build rule.

Since: 0.1.0

makeSpecialDepsMSVC :: Text -> SpecialDeps

Construct a SpecialDeps corresponding to the case in which deps = msvc is set and msvc_deps_prefix = ….

The msvc_deps_prefix field defines the string which should be stripped from msvc's /showIncludes output. It is only needed if the version of Visual Studio being used is not English. The value of msvc_deps_prefix is "Note: including file: " by default.

Since: 0.1.0

_SpecialDepsGCC :: Prism' SpecialDeps ()

A prism for the deps = gcc case.

Since: 0.1.0

_SpecialDepsMSVC :: Prism' SpecialDeps Text

A prism for the deps = msvc / msvc_deps_prefix = … case.

Since: 0.1.0

data ResponseFile

A response file to use during rule execution, as documented here.

Since: 0.1.0

makeResponseFile

Arguments

:: Path
Corresponds to rspfile.

-> Text
Corresponds to rspfile_content.

-> ResponseFile

Construct a ResponseFile with the given Path and content Text.

Since: 0.1.0

responseFilePath :: Lens' ResponseFile Path

A lens for the rspfile field.

Since: 0.1.0

responseFileContent :: Lens' ResponseFile Text

A lens for the rspfile_content field.

Since: 0.1.0

Language.Ninja.IR.Target

data Target

This type represents a Ninja target name.

Since: 0.1.0

makeTarget :: Text -> Target

Construct a Target from some Text.

Since: 0.1.0

targetIText :: Iso' Target IText

An isomorphism between a Target and its underlying IText.

Since: 0.1.0

targetText :: Iso' Target Text

An isomorphism that gives access to a Text-typed view of a Target, even though the underlying data has type IText.

This is equivalent to targetIText . from itext.

Since: 0.1.0

data Output

A Ninja build output.

More information is available here.

Since: 0.1.0

makeOutput

Arguments

:: Target
The underlying target.

-> Output

Construct an Output.

Since: 0.1.0

outputTarget :: Lens' Output Target

A lens for the Target of an Output.

Since: 0.1.0

data Dependency

A build dependency.

More information is available here.

Since: 0.1.0

makeDependency

Arguments

:: Target
The underlying target.

-> DependencyType
The dependency type (normal, implicit, or order-only).

-> Dependency

Construct a Dependency.

Since: 0.1.0

dependencyTarget :: Lens' Dependency Target

A lens for the Target of a Dependency.

Since: 0.1.0

dependencyType :: Lens' Dependency DependencyType

A lens for the DependencyType of a Dependency.

Since: 0.1.0

data DependencyType

The type of a Dependency: normal, implicit, or order-only.

Since: 0.1.0

Constructors

NormalDependency
A normal dependency. These are listed in the $in variable and changes in the relevant target result in a rule execution.

Since: 0.1.0

OrderOnlyDependency
An order-only dependency. These are listed in the $in variable, but are only rebuilt if there is at least one non-order-only dependency that is out of date.

FIXME: double check this interpretation of the Ninja manual

Since: 0.1.0

_NormalDependency :: Prism' DependencyType ()

A prism for the NormalDependency constructor.

Since: 0.1.0

_OrderOnlyDependency :: Prism' DependencyType ()

A prism for the OrderOnlyDependency constructor.

Since: 0.1.0

Language.Ninja.IR.Pool

data Pool

A Ninja pool declaration, as documented here.

Since: 0.1.0

makePool :: PoolName -> PoolDepth -> Maybe Pool

Construct a Pool, given its name and depth.

Since: 0.1.0

makePoolDefault :: Pool

The default pool, i.e.: the one whose name is the empty string.

Since: 0.1.0

makePoolConsole :: Pool

The console pool.

Since: 0.1.0

makePoolCustom

Arguments

:: Text
The pool name.

-> Positive
The pool depth.

-> Pool

Create a pool with the given name and depth.

Since: 0.1.0

poolName :: Getter Pool PoolName

A Getter that gives the name of a pool.

Since: 0.1.0

poolDepth :: Getter Pool PoolDepth

A Getter that gives the depth of a pool.

Since: 0.1.0

data PoolName

The name of a Ninja pool.

More information is available here.

Since: 0.1.0

makePoolNameDefault :: PoolName

Create a PoolName corresponding to the built-in default pool, i.e.: the pool that is selected if the pool attribute is set to the empty string.

Since: 0.1.0

makePoolNameConsole :: PoolName

Create a PoolName corresponding to the built-in console pool.

Since: 0.1.0

makePoolNameCustom :: Text -> PoolName

Create a PoolName corresponding to a custom pool. Note: this can fail at runtime if given the empty string or "console", so you should consider parsePoolName as a safer alternative.

Since: 0.1.0

_PoolNameDefault :: Getter PoolName (Maybe ())

A one-way prism corresponding to the poolNameDefault constructor.

Since: 0.1.0

_PoolNameConsole :: Getter PoolName (Maybe ())

A one-way prism corresponding to the poolNameConsole constructor.

Since: 0.1.0

_PoolNameCustom :: Getter PoolName (Maybe Text)

A one-way prism corresponding to the poolNameConsole constructor.

Since: 0.1.0

poolNameText :: Iso' PoolName Text

An isomorphism between a PoolName and the corresponding Text. Equivalent to iso printPoolName parsePoolName.

Since: 0.1.0

printPoolName :: PoolName -> Text

Convert a PoolName to the string that, if the pool attribute is set to it, will cause the given PoolName to be parsed.

>>> printPoolName poolNameDefault
""

>>> printPoolName poolNameConsole
"console"

>>> printPoolName (poolNameCustom "foobar")
"foobar"

Since: 0.1.0

parsePoolName :: Text -> PoolName

Inverse of printPoolName.

>>> parsePoolName ""
PoolNameDefault

>>> parsePoolName "console"
PoolNameConsole

>>> parsePoolName "foobar"
PoolNameCustom "foobar"

Since: 0.1.0

data PoolDepth

The depth of a Ninja pool.

More information is available here.

Since: 0.1.0

makePoolDepth :: Positive -> PoolDepth

Construct a finite PoolDepth from an integer, which should be a number greater than or equal to 1.

Since: 0.1.0

makePoolInfinite :: PoolDepth

Construct an infinite PoolDepth. This constructor is needed for the default pool (pool = ""), which has an infinite depth.

Since: 0.1.0

poolDepthPositive :: Iso' PoolDepth (Maybe Positive)

An isomorphism between a PoolDepth and a Maybe Positive; the Nothing case maps to makePoolInfinite and the Just case maps to makePoolDepth.

Since: 0.1.0