Language.Ninja.IR.Build

Description

A datatype for Ninja build declarations.

Since: 0.1.0

Build

data Build

A Ninja build declaration, as documented here.

Since: 0.1.0

makeBuild :: Rule -> Build

Construct a default Build from the given Rule

Since: 0.1.0

buildRule :: Lens' Build Rule

The rule to execute when building any of the outputs.

Since: 0.1.0

buildOuts :: Lens' Build (HashSet Output)

The outputs that are built as a result of rule execution.

Since: 0.1.0

buildDeps :: Lens' Build (HashSet Dependency)

The dependencies that must be satisfied before this can be built.

Since: 0.1.0

type BuildConstraint c = (c Text, c (HashSet Output), c (HashSet Dependency))

The set of constraints required for a given constraint to be automatically computed for a Build.

Since: 0.1.0