Language.Ninja.IR.Ninja

Description

A datatype representing the intermediate representation of a Ninja file after compilation.

Since: 0.1.0

Ninja

data Ninja

A parsed and normalized Ninja file.

Since: 0.1.0

makeNinja :: Ninja

Construct a default Ninja value.

Since: 0.1.0

ninjaMeta :: Lens' Ninja Meta

Metadata, which includes top-level variables like builddir.

Since: 0.1.0

ninjaBuilds :: Lens' Ninja (HashSet Build)

Compiled build declarations.

Since: 0.1.0

ninjaPhonys :: Lens' Ninja (HashMap Target (HashSet Target))

Phony targets, as documented here.

Since: 0.1.0

ninjaDefaults :: Lens' Ninja (HashSet Target)

The set of default targets, as documented here.

Since: 0.1.0

ninjaPools :: Lens' Ninja (HashSet Pool)

The set of pools for this Ninja file.

Since: 0.1.0

type NinjaConstraint c = (c Text, c Version, c (HashMap Target (HashSet Target)), c (HashSet Build), c (HashSet Target), c (HashSet Pool))

The set of constraints required for a given constraint to be automatically computed for a Ninja.

Since: 0.1.0