Language.Ninja.IR.Pool

Description

Types relating to Ninja pools.

Since: 0.1.0

Pool

data Pool

A Ninja pool declaration, as documented here.

Since: 0.1.0

makePool :: PoolName -> PoolDepth -> Maybe Pool

Construct a Pool, given its name and depth.

Since: 0.1.0

makePoolDefault :: Pool

The default pool, i.e.: the one whose name is the empty string.

Since: 0.1.0

makePoolConsole :: Pool

The console pool.

Since: 0.1.0

makePoolCustom

Arguments

:: Text
The pool name.

-> Positive
The pool depth.

-> Pool

Create a pool with the given name and depth.

Since: 0.1.0

poolName :: Getter Pool PoolName

A Getter that gives the name of a pool.

Since: 0.1.0

poolDepth :: Getter Pool PoolDepth

A Getter that gives the depth of a pool.

Since: 0.1.0

PoolName

data PoolName

The name of a Ninja pool.

More information is available here.

Since: 0.1.0

makePoolNameDefault :: PoolName

Create a PoolName corresponding to the built-in default pool, i.e.: the pool that is selected if the pool attribute is set to the empty string.

Since: 0.1.0

makePoolNameConsole :: PoolName

Create a PoolName corresponding to the built-in console pool.

Since: 0.1.0

makePoolNameCustom :: Text -> PoolName

Create a PoolName corresponding to a custom pool. Note: this can fail at runtime if given the empty string or "console", so you should consider parsePoolName as a safer alternative.

Since: 0.1.0

_PoolNameDefault :: Getter PoolName (Maybe ())

A one-way prism corresponding to the poolNameDefault constructor.

Since: 0.1.0

_PoolNameConsole :: Getter PoolName (Maybe ())

A one-way prism corresponding to the poolNameConsole constructor.

Since: 0.1.0

_PoolNameCustom :: Getter PoolName (Maybe Text)

A one-way prism corresponding to the poolNameConsole constructor.

Since: 0.1.0

poolNameText :: Iso' PoolName Text

An isomorphism between a PoolName and the corresponding Text. Equivalent to iso printPoolName parsePoolName.

Since: 0.1.0

printPoolName :: PoolName -> Text

Convert a PoolName to the string that, if the pool attribute is set to it, will cause the given PoolName to be parsed.

>>> printPoolName poolNameDefault
""

>>> printPoolName poolNameConsole
"console"

>>> printPoolName (poolNameCustom "foobar")
"foobar"

Since: 0.1.0

parsePoolName :: Text -> PoolName

Inverse of printPoolName.

>>> parsePoolName ""
PoolNameDefault

>>> parsePoolName "console"
PoolNameConsole

>>> parsePoolName "foobar"
PoolNameCustom "foobar"

Since: 0.1.0

PoolDepth

data PoolDepth

The depth of a Ninja pool.

More information is available here.

Since: 0.1.0

makePoolDepth :: Positive -> PoolDepth

Construct a finite PoolDepth from an integer, which should be a number greater than or equal to 1.

Since: 0.1.0

makePoolInfinite :: PoolDepth

Construct an infinite PoolDepth. This constructor is needed for the default pool (pool = ""), which has an infinite depth.

Since: 0.1.0

poolDepthPositive :: Iso' PoolDepth (Maybe Positive)

An isomorphism between a PoolDepth and a Maybe Positive; the Nothing case maps to makePoolInfinite and the Just case maps to makePoolDepth.

Since: 0.1.0