Language.Ninja.IR.Rule

Description

A datatype for Ninja rule declarations.

Since: 0.1.0

Rule

data Rule

A Ninja rule declaration, as documented here.

Since: 0.1.0

makeRule

Arguments

:: Text
The rule name.

-> Command
The command to run.

-> Rule
A rule that runs this command.

Construct an Rule with the given name and command, with default values for all other attributes (e.g.: False, Nothing, poolDefault).

Since: 0.1.0

ruleName :: Lens' Rule Text

The name of the rule.

Since: 0.1.0

ruleCommand :: Lens' Rule Command

The command that this rule will run.

Since: 0.1.0

ruleDescription :: Lens' Rule (Maybe Text)

A short description of the command, used to pretty-print the command as it's running. The ninja -v flag controls whether to print the full command or its description; if a command fails, the full command line will always be printed before the command's output.

Since: 0.1.0

rulePool :: Lens' Rule PoolName

The process pool in which this rule will be executed.

Since: 0.1.0

ruleDepfile :: Lens' Rule (Maybe Path)

If set, this should be a path to an optional Makefile that contains extra implicit dependencies. This is used to support C/C++ header dependencies. For more information, read the Ninja documentation here.

Since: 0.1.0

ruleSpecialDeps :: Lens' Rule (Maybe SpecialDeps)

If set, enables special dependency processing used in C/C++ header dependencies. For more information, read the Ninja documentation here.

Since: 0.1.0

ruleGenerator :: Lens' Rule Bool

If this is true, specifies that this rule is used to re-invoke the generator program. Files built using generator rules are treated specially in two ways: firstly, they will not be rebuilt if the command line changes; and secondly, they are not cleaned by default.

Since: 0.1.0

ruleRestat :: Lens' Rule Bool

If true, causes Ninja to re-stat the command's outputs after execution of the command. Each output whose modification time the command did not change will be treated as though it had never needed to be built. This may cause the output's reverse dependencies to be removed from the list of pending build actions.

Since: 0.1.0

ruleResponseFile :: Lens' Rule (Maybe ResponseFile)

If present, Ninja will use a response file for the given command, i.e. write the selected string to the given file before calling the command and delete the file after the command is done.

This is particularly useful on Windows OS, where the maximal length of a command line is limited and response files must be used instead.

Since: 0.1.0

SpecialDeps

data SpecialDeps

Special dependency information, as described here.

Since: 0.1.0

makeSpecialDepsGCC :: SpecialDeps

Construct a SpecialDeps corresponding to the case in which deps = gcc is set in a Ninja build rule.

Since: 0.1.0

makeSpecialDepsMSVC :: Text -> SpecialDeps

Construct a SpecialDeps corresponding to the case in which deps = msvc is set and msvc_deps_prefix = ….

The msvc_deps_prefix field defines the string which should be stripped from msvc's /showIncludes output. It is only needed if the version of Visual Studio being used is not English. The value of msvc_deps_prefix is "Note: including file: " by default.

Since: 0.1.0

_SpecialDepsGCC :: Prism' SpecialDeps ()

A prism for the deps = gcc case.

Since: 0.1.0

_SpecialDepsMSVC :: Prism' SpecialDeps Text

A prism for the deps = msvc / msvc_deps_prefix = … case.

Since: 0.1.0

ResponseFile

data ResponseFile

A response file to use during rule execution, as documented here.

Since: 0.1.0

makeResponseFile

Arguments

:: Path
Corresponds to rspfile.

-> Text
Corresponds to rspfile_content.

-> ResponseFile

Construct a ResponseFile with the given Path and content Text.

Since: 0.1.0

responseFilePath :: Lens' ResponseFile Path

A lens for the rspfile field.

Since: 0.1.0

responseFileContent :: Lens' ResponseFile Text

A lens for the rspfile_content field.

Since: 0.1.0