Language.Ninja.IR.Target

Description

Types relating to Ninja build targets, outputs, and dependencies.

Since: 0.1.0

Target

data Target

This type represents a Ninja target name.

Since: 0.1.0

makeTarget :: Text -> Target

Construct a Target from some Text.

Since: 0.1.0

targetIText :: Iso' Target IText

An isomorphism between a Target and its underlying IText.

Since: 0.1.0

targetText :: Iso' Target Text

An isomorphism that gives access to a Text-typed view of a Target, even though the underlying data has type IText.

This is equivalent to targetIText . from itext.

Since: 0.1.0

Output

data Output

A Ninja build output.

More information is available here.

Since: 0.1.0

makeOutput

Arguments

:: Target
The underlying target.

-> Output

Construct an Output.

Since: 0.1.0

outputTarget :: Lens' Output Target

A lens for the Target of an Output.

Since: 0.1.0

Dependency

data Dependency

A build dependency.

More information is available here.

Since: 0.1.0

makeDependency

Arguments

:: Target
The underlying target.

-> DependencyType
The dependency type (normal, implicit, or order-only).

-> Dependency

Construct a Dependency.

Since: 0.1.0

dependencyTarget :: Lens' Dependency Target

A lens for the Target of a Dependency.

Since: 0.1.0

dependencyType :: Lens' Dependency DependencyType

A lens for the DependencyType of a Dependency.

Since: 0.1.0

data DependencyType

The type of a Dependency: normal, implicit, or order-only.

Since: 0.1.0

Constructors

NormalDependency
A normal dependency. These are listed in the $in variable and changes in the relevant target result in a rule execution.

Since: 0.1.0

OrderOnlyDependency
An order-only dependency. These are listed in the $in variable, but are only rebuilt if there is at least one non-order-only dependency that is out of date.

FIXME: double check this interpretation of the Ninja manual

Since: 0.1.0

_NormalDependency :: Prism' DependencyType ()

A prism for the NormalDependency constructor.

Since: 0.1.0

_OrderOnlyDependency :: Prism' DependencyType ()

A prism for the OrderOnlyDependency constructor.

Since: 0.1.0