Language.Ninja.Lexer

Description

Lexing is a slow point, the code below is optimised.

Since: 0.1.0

Lexing

lexerFile :: (MonadError ParseError m, MonadReadFile m) => Path -> m [Lexeme Ann]

Lex the given file.

Since: 0.1.0

lexerText :: MonadError ParseError m => Text -> m [Lexeme Ann]

Lex the given Text.

Since: 0.1.0

lexerBS :: MonadError ParseError m => ByteString -> m [Lexeme Ann]

Lex the given ByteString.

Since: 0.1.0

lexerText' :: MonadError ParseError m => Maybe Path -> Text -> m [Lexeme Ann]

Lex the given Text that comes from the given Path, if provided.

Since: 0.1.0

lexerBS' :: MonadError ParseError m => Maybe Path -> ByteString -> m [Lexeme Ann]

Lex the given ByteString that comes from the given Path, if it is provided. The Path is only used for error messages.

Since: 0.1.0

lexemesP :: Parser m [Lexeme Ann]

The megaparsec parser for a Ninja file.

Since: 0.1.0

type Parser m a = ParsecT Dec Text m a

A megaparsec parser.

Since: 0.1.0

type Ann = Spans

The type of annotations returned by the lexer.

Since: 0.1.0

Lexeme and friends

data Lexeme ann

Lex each line separately, rather than each lexeme.

Since: 0.1.0

Constructors

LexDefine !ann !(LBind ann)
foo = bar

Since: 0.1.0

LexBind !ann !(LBind ann)
[indent]foo = bar

Since: 0.1.0

LexInclude !ann !(LFile ann)
include file

Since: 0.1.0

LexSubninja !ann !(LFile ann)
subninja file

Since: 0.1.0

LexBuild !ann !(LBuild ann)
build foo: bar | baz || qux

Since: 0.1.0

LexRule !ann !(LName ann)
rule name

Since: 0.1.0

LexPool !ann !(LName ann)
pool name

Since: 0.1.0

LexDefault !ann ![Expr ann]
default foo bar

Since: 0.1.0

data LName ann

The name of a Ninja rule or pool.

Since: 0.1.0

Constructors

MkLName

Fields

_lnameAnn :: !ann

_lnameBS :: !ByteString

newtype LFile ann

A reference to a file in an include or subninja declaration.

Since: 0.1.0

Constructors

MkLFile

Fields

_lfileExpr :: Expr ann

data LBind ann

A Ninja variable binding, top-level or otherwise.

Since: 0.1.0

Constructors

MkLBind

Fields

_lbindAnn :: !ann

_lbindName :: !(LName ann)

_lbindValue :: !(Expr ann)

data LBuild ann

The data contained within a Ninja build declaration.

Since: 0.1.0

Constructors

MkLBuild

Fields

_lbuildAnn :: !ann

_lbuildOuts :: ![Expr ann]

_lbuildRule :: !(LName ann)

_lbuildDeps :: ![Expr ann]

makeLBuild

Arguments

:: ann
The build annotation

-> [Expr ann]
The build outputs

-> LName ann
The rule name

-> [Expr ann]
The build dependencies

-> LBuild ann

Constructor for an LBuild.

Since: 0.1.0

type LexemeConstraint c ann = (LBindConstraint c ann, LFileConstraint c ann, LBuildConstraint c ann, LNameConstraint c ann, c [Expr ann], c ann)

The set of constraints required for a given constraint to be automatically computed for an Lexeme.

Since: 0.1.0

type LNameConstraint c ann = (c Text, c ann)

The set of constraints required for a given constraint to be automatically computed for an LName.

Since: 0.1.0

type LFileConstraint c ann = (c Text, c ann)

The set of constraints required for a given constraint to be automatically computed for an LFile.

Since: 0.1.0

type LBindConstraint c ann = (c Text, c ann)

The set of constraints required for a given constraint to be automatically computed for an LBind.

Since: 0.1.0

type LBuildConstraint c ann = (c Text, c ann)

The set of constraints required for a given constraint to be automatically computed for an LBuild.

Since: 0.1.0

Classes

class Monad m => PositionParsing m where

This class is kind of like DeltaParsing from trifecta.

Since: 0.1.0

Minimal complete definition

getPosition

Methods

getPosition :: m Position