Language.Ninja.Misc

Description

Utility types and functions for the rest of language-ninja.

This module re-exports all of the modules under the Language.Ninja.Misc namespace for convenience.

It is recommended that you import it with the following style:

import qualified Language.Ninja.Misc as Misc

Since: 0.1.0

Language.Ninja.Misc.Command

data Command

This type represents a POSIX sh command line.

Since: 0.1.0

makeCommand :: Text -> Command

Constructor for a Command.

Since: 0.1.0

commandText :: Iso' Command Text

An isomorphism between a Command and its underlying Text.

Since: 0.1.0

Language.Ninja.Misc.Path

data Path

This type represents a Unix path string.

Since: 0.1.0

makePath :: Text -> Path

Construct a Path from some Text.

Since: 0.1.0

pathIText :: Iso' Path IText

An isomorphism between a Path and its underlying IText.

Since: 0.1.0

pathText :: Iso' Path Text

An isomorphism that gives access to a Text-typed view of a Path, even though the underlying data has type IText.

This is equivalent to pathIText . from Ninja.itext.

Since: 0.1.0

pathString :: Iso' Path String

An isomorphism that gives access to a String-typed view of a Path.

Since: 0.1.0

pathFP :: Iso' Path FilePath

An isomorphism between a Path and a FilePath from system-filepath. This uses decodeString and encodeString, so all the caveats on those functions apply here.

Since: 0.1.0

Language.Ninja.Misc.IText

data IText

An interned (hash-consed) text type. This is a newtype over InternedText from the intern package.

Since: 0.1.0

uninternText :: IText -> Text

Get the Text corresponding to the given IText value.

internText (uninternText x) = x

>>> uninternText ("foobar" :: IText)
"foobar"

Since: 0.1.0

internText :: Text -> IText

Intern a Text value, resulting in an IText value.

uninternText (internText x) = x

>>> internText ("foobar" :: Text)
"foobar"

Since: 0.1.0

itext :: Iso' Text IText

An Iso' between Text and IText.

((fromString x) ^. itext) = T.pack x

>>> (("foobar" :: Text) ^. itext) :: IText
"foobar"

>>> (("foobar" :: IText) ^. from itext) :: Text
"foobar"

Since: 0.1.0

Language.Ninja.Misc.Positive

data Positive

This type represents a positive number; i.e.: an integer greater than zero.

Since: 0.1.0

makePositive :: Int -> Maybe Positive

Constructor for a Positive.

Since: 0.1.0

fromPositive :: Getter Positive Int

A Getter for the Int underlying a Positive.

Since: 0.1.0

Language.Ninja.Misc.Annotated

class Functor ty => Annotated ty where

If you have some type that represents an AST node, it is often useful to add a polymorphic "annotation field" to it, which is used for things like source positions.

Specifically, suppose we have the following AST node type:

data Foo = Foo { _fooBar :: !Bar, _fooBaz :: !Baz } deriving (…)

Then an annotation field is added by the following process:

Add an extra (final) type parameter ann to the type.

Add an extra field _fooAnn :: !ann.

Derive instances of Functor, Foldable, and Traversable.

If the type is recursive, add a Plated instance. See Language.Ninja.AST.Expr for a complete example of this.

Write an Annotated instance with the canonical lens given by the _fooAnn field. There are plenty of examples around this library.

The end result then looks like:

data Foo ann
  = Foo
    { _fooAnn :: !ann
    , _fooBar :: !Bar
    , _fooBaz :: !Baz
    }
  deriving (…, Functor, Foldable, Traversable)

instance Annotated Foo where
  annotation' = …

Since: 0.1.0

Minimal complete definition

annotation'

Methods

annotation' :: (ann -> ann') -> Lens (ty ann) (ty ann') ann ann'

Given a function that is used when fmaping any subterms, return a lens into the "annotation" field.

When writing an instance, keep in mind that annotation' id should just be the typical definition for a lens into the annotation field.

annotation' (f . g) == annotation' f . annotation' g

Since: 0.1.0

annotation :: Annotated ty => Lens' (ty ann) ann

This is just shorthand for annotation' id.

Since: 0.1.0

Language.Ninja.Misc.Located

data Located t

This datatype represents a value annotated with a source location.

Since: 0.1.0

tokenize :: Maybe Path -> Text -> [Located Text]

Given path :: Maybe Path and a text :: Text, do the following:

Remove all '\r' characters from the text.

Split the text into chunks that are guaranteed not to contain newlines or whitespace, and which are annotated with their location.

Since: 0.1.0

tokenizeFile :: MonadReadFile m => Path -> m [Located Text]

Read the file at the given Path and then run tokenize on the resulting Text.

Since: 0.1.0

tokenizeText :: Text -> [Located Text]

This function is equivalent to tokenize Nothing.

Since: 0.1.0

locatedPos :: Lens' (Located t) Position

The position of this located value.

Since: 0.1.0

locatedVal :: Lens' (Located t) t

The value underlying this located value.

Since: 0.1.0

data Spans

A type representing a set of source spans.

Since: 0.1.0

makeSpans :: [Span] -> Spans

Construct a Spans from a list of Spans.

Since: 0.1.0

spansSet :: Iso' Spans (HashSet Span)

A lens into the HashSet Span underlying a value of type Spans.

Since: 0.1.0

data Span

Represents a span of source code.

Since: 0.1.0

makeSpan

Arguments

:: Maybe Path
The file in which this span resides, if any.

-> Offset
The start offset.

-> Offset
The end offset.

-> Span

Construct a Span from a given start position to a given end position.

Since: 0.1.0

spanPath :: Lens' Span (Maybe Path)

A lens into the (nullable) path associated with a Span.

Since: 0.1.0

spanRange :: Lens' Span (Offset, Offset)

A lens giving the start and end Offsets associated with a Span.

Since: 0.1.0

spanStart :: Lens' Span Offset

A lens into the Offset associated with the start of a Span.

Since: 0.1.0

spanEnd :: Lens' Span Offset

A lens into the Offset associated with the end of a Span.

Since: 0.1.0

data Position

This datatype represents the position of a cursor in a text file.

Since: 0.1.0

makePosition :: Maybe Path -> Offset -> Position

Construct a Position from a (nullable) path and a (line, column) pair.

Since: 0.1.0

positionFile :: Lens' Position (Maybe Path)

The path of the file pointed to by this position, if any.

Since: 0.1.0

positionOffset :: Lens' Position Offset

The offset in the file pointed to by this position.

Since: 0.1.0

positionLine :: Lens' Position Line

The line number in the file pointed to by this position.

Since: 0.1.0

positionCol :: Lens' Position Column

The column number in the line pointed to by this position.

Since: 0.1.0

comparePosition :: Position -> Position -> Maybe Ordering

If two Positions are comparable (i.e.: if they are in the same file), this function will return an Ordering giving their relative positions. Otherwise, it will of course return Nothing.

Since: 0.1.0

type Offset = (Line, Column)

A line/column offset into a file.

Since: 0.1.0

compareOffset :: Offset -> Offset -> Ordering

Compare two Offsets in lexicographic order (i.e.: the Column is ignored unless they are on the same Line).

Since: 0.1.0

offsetLine :: Lens' Offset Line

A lens into the Line associated with an Offset.

For now, this is simply defined as offsetLine = _1, but if Offset is later refactored to be an abstract data type, using this lens instead of _1 will decrease the amount of code that breaks.

Since: 0.1.0

offsetColumn :: Lens' Offset Column

A lens into the Line associated with an Offset.

Read the description of offsetLine for an understanding of why this exists and why you should use it instead of _2.

Since: 0.1.0

type Line = Int

A line number.

Since: 0.1.0

type Column = Int

A column number.

Since: 0.1.0