Language.Ninja.Misc.Annotated

Description

A typeclass for AST nodes that are annotated with a polymorphic field, which provides a canonical lens into that field.

Since: 0.1.0

Documentation

class Functor ty => Annotated ty where

If you have some type that represents an AST node, it is often useful to add a polymorphic "annotation field" to it, which is used for things like source positions.

Specifically, suppose we have the following AST node type:

data Foo = Foo { _fooBar :: !Bar, _fooBaz :: !Baz } deriving (…)

Then an annotation field is added by the following process:

Add an extra (final) type parameter ann to the type.

Add an extra field _fooAnn :: !ann.

Derive instances of Functor, Foldable, and Traversable.

If the type is recursive, add a Plated instance. See Language.Ninja.AST.Expr for a complete example of this.

Write an Annotated instance with the canonical lens given by the _fooAnn field. There are plenty of examples around this library.

The end result then looks like:

data Foo ann
  = Foo
    { _fooAnn :: !ann
    , _fooBar :: !Bar
    , _fooBaz :: !Baz
    }
  deriving (…, Functor, Foldable, Traversable)

instance Annotated Foo where
  annotation' = …

Since: 0.1.0

Minimal complete definition

annotation'

Methods

annotation' :: (ann -> ann') -> Lens (ty ann) (ty ann') ann ann'

Given a function that is used when fmaping any subterms, return a lens into the "annotation" field.

When writing an instance, keep in mind that annotation' id should just be the typical definition for a lens into the annotation field.

annotation' (f . g) == annotation' f . annotation' g

Since: 0.1.0

annotation :: Annotated ty => Lens' (ty ann) ann

This is just shorthand for annotation' id.

Since: 0.1.0