Language.Ninja.Misc.Located

Description

Tokenize text into a list of non-whitespace chunks, each of which is annotated with its source location.

Since: 0.1.0

Located

data Located t

This datatype represents a value annotated with a source location.

Since: 0.1.0

tokenize :: Maybe Path -> Text -> [Located Text]

Given path :: Maybe Path and a text :: Text, do the following:

Remove all '\r' characters from the text.

Split the text into chunks that are guaranteed not to contain newlines or whitespace, and which are annotated with their location.

Since: 0.1.0

tokenizeFile :: MonadReadFile m => Path -> m [Located Text]

Read the file at the given Path and then run tokenize on the resulting Text.

Since: 0.1.0

tokenizeText :: Text -> [Located Text]

This function is equivalent to tokenize Nothing.

Since: 0.1.0

untokenize :: [Located Text] -> Map Path Text

This function takes the output of tokenize and returns a map from paths to the contents of the associated files.

untokenize [] == Map.empty

untokenize (xs <> ys) == untokenize xs <> untokenize ys

untokenize (tokenize (Just path) t) == Map.singleton path t

FIXME: implement

Since: 0.1.0

locatedPos :: Lens' (Located t) Position

The position of this located value.

Since: 0.1.0

locatedVal :: Lens' (Located t) t

The value underlying this located value.

Since: 0.1.0

Spans

data Spans

A type representing a set of source spans.

Since: 0.1.0

makeSpans :: [Span] -> Spans

Construct a Spans from a list of Spans.

Since: 0.1.0

spansSet :: Iso' Spans (HashSet Span)

A lens into the HashSet Span underlying a value of type Spans.

Since: 0.1.0

Span

data Span

Represents a span of source code.

Since: 0.1.0

makeSpan

Arguments

:: Maybe Path
The file in which this span resides, if any.

-> Offset
The start offset.

-> Offset
The end offset.

-> Span

Construct a Span from a given start position to a given end position.

Since: 0.1.0

spanPath :: Lens' Span (Maybe Path)

A lens into the (nullable) path associated with a Span.

Since: 0.1.0

spanRange :: Lens' Span (Offset, Offset)

A lens giving the start and end Offsets associated with a Span.

Since: 0.1.0

spanStart :: Lens' Span Offset

A lens into the Offset associated with the start of a Span.

Since: 0.1.0

spanEnd :: Lens' Span Offset

A lens into the Offset associated with the end of a Span.

Since: 0.1.0

spanStartPos :: Getter Span Position

A getter for the Position associated with the start of a Span.

Since: 0.1.0

spanEndPos :: Getter Span Position

A getter for the Position associated with the end of a Span.

Since: 0.1.0

Position

data Position

This datatype represents the position of a cursor in a text file.

Since: 0.1.0

makePosition :: Maybe Path -> Offset -> Position

Construct a Position from a (nullable) path and a (line, column) pair.

Since: 0.1.0

positionFile :: Lens' Position (Maybe Path)

The path of the file pointed to by this position, if any.

Since: 0.1.0

positionOffset :: Lens' Position Offset

The offset in the file pointed to by this position.

Since: 0.1.0

positionLine :: Lens' Position Line

The line number in the file pointed to by this position.

Since: 0.1.0

positionCol :: Lens' Position Column

The column number in the line pointed to by this position.

Since: 0.1.0

comparePosition :: Position -> Position -> Maybe Ordering

If two Positions are comparable (i.e.: if they are in the same file), this function will return an Ordering giving their relative positions. Otherwise, it will of course return Nothing.

Since: 0.1.0

Offset

type Offset = (Line, Column)

A line/column offset into a file.

Since: 0.1.0

compareOffset :: Offset -> Offset -> Ordering

Compare two Offsets in lexicographic order (i.e.: the Column is ignored unless they are on the same Line).

Since: 0.1.0

offsetLine :: Lens' Offset Line

A lens into the Line associated with an Offset.

For now, this is simply defined as offsetLine = _1, but if Offset is later refactored to be an abstract data type, using this lens instead of _1 will decrease the amount of code that breaks.

Since: 0.1.0

offsetColumn :: Lens' Offset Column

A lens into the Line associated with an Offset.

Read the description of offsetLine for an understanding of why this exists and why you should use it instead of _2.

Since: 0.1.0

Miscellaneous

type Line = Int

A line number.

Since: 0.1.0

type Column = Int

A column number.

Since: 0.1.0