Language.Ninja.Misc.Path

Description

A datatype for Unix path strings.

Since: 0.1.0

Documentation

data Path

This type represents a Unix path string.

Since: 0.1.0

makePath :: Text -> Path

Construct a Path from some Text.

Since: 0.1.0

pathIText :: Iso' Path IText

An isomorphism between a Path and its underlying IText.

Since: 0.1.0

pathText :: Iso' Path Text

An isomorphism that gives access to a Text-typed view of a Path, even though the underlying data has type IText.

This is equivalent to pathIText . from Ninja.itext.

Since: 0.1.0

pathString :: Iso' Path String

An isomorphism that gives access to a String-typed view of a Path.

Since: 0.1.0

pathFP :: Iso' Path FilePath

An isomorphism between a Path and a FilePath from system-filepath. This uses decodeString and encodeString, so all the caveats on those functions apply here.

Since: 0.1.0