Language.Ninja.Mock.ReadFile

Description

A typeclass exposing an interface that allows reading a file from a filesystem (virtual or otherwise).

Since: 0.1.0

Documentation

class Monad m => MonadReadFile m where

This typeclass allows you to write code that reads files from the filesystem and then later run that code purely against a virtual filesystem of some description.

Since: 0.1.0

Minimal complete definition

readFile

Methods

readFile :: Path -> m Text

Read the file located at the given path and decode it into Text.

FIXME: some notion of error handling should be encoded into the type

Since: 0.1.0