Language.Ninja.Parser

Description

Parse a Ninja file.

Since: 0.1.0

parse*IO

parseFileIO :: Path -> IO (Ninja Ann)

Parse the file at the given path into a Ninja. This function may throw an exception if parsing fails.

Since: 0.1.0

parseTextIO :: Text -> IO (Ninja Ann)

Parse the given Text into a Ninja. This function may throw an exception if parsing fails.

Since: 0.1.0

parseBSIO :: ByteString -> IO (Ninja Ann)

Parse the given ByteString into a Ninja. This function may throw an exception if parsing fails.

Since: 0.1.0

parseLexemesIO :: [Lexeme Ann] -> IO (Ninja Ann)

Parse the given list of Lexemes into a Ninja. This function may throw an exception if parsing fails.

Since: 0.1.0

parse*

parseFile :: (MonadError ParseError m, MonadReadFile m) => Path -> m (Ninja Ann)

Parse the file at the given path into a Ninja.

Since: 0.1.0

parseText :: (MonadError ParseError m, MonadReadFile m) => Text -> m (Ninja Ann)

Parse the given Text into a Ninja.

Since: 0.1.0

parseBS :: (MonadError ParseError m, MonadReadFile m) => ByteString -> m (Ninja Ann)

Parse the given ByteString into a Ninja.

Since: 0.1.0

parseLexemes :: (MonadError ParseError m, MonadReadFile m) => [Lexeme Ann] -> m (Ninja Ann)

Parse the given list of Lexemes into a Ninja.

Since: 0.1.0

parse*WithEnv

parseFileWithEnv :: (MonadError ParseError m, MonadReadFile m) => Path -> Env Text Text -> m (Ninja Ann)

Parse the file at the given path using the given Ninja variable context, resulting in a Ninja.

Since: 0.1.0

parseTextWithEnv :: (MonadError ParseError m, MonadReadFile m) => Text -> Env Text Text -> m (Ninja Ann)

Parse the given Text using the given Ninja variable context, resulting in a Ninja.

Since: 0.1.0

parseBSWithEnv :: (MonadError ParseError m, MonadReadFile m) => ByteString -> Env Text Text -> m (Ninja Ann)

Parse the given ByteString using the given Ninja variable context, resulting in a Ninja.

Since: 0.1.0

parseLexemesWithEnv :: (MonadError ParseError m, MonadReadFile m) => [Lexeme Ann] -> Env Text Text -> m (Ninja Ann)

Parse the given list of Lexemes into a Ninja.

Since: 0.1.0